requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bce-python-sdk==0.8.83
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==41.0.0
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.35
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.0
google-auth-oauthlib==1.0.0
gradio==3.11.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
hydra-core==1.3.2
idna==3.4
imageio==2.30.0
imgaug==0.4.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
linkify-it-py==2.0.2
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.12.0
opencv-contrib-python
opencv-python
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.14
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleocr
paddlepaddle
pandas==2.0.2
paramiko==3.2.0
parso==0.8.3
pdf2docx==0.5.6
pickleshare==0.7.5
Pillow==9.5.0
premailer==3.10.0
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
PyMuPDF==1.20.2
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.0.0
rarfile==4.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy
seaborn==0.12.2
sentry-sdk==1.24.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
tritonclient==2.34.0
typing_extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
ultralytics==8.0.104
urllib3==1.26.16
uvicorn==0.22.0
visualdl==2.5.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
x2paddle==1.4.1
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
